Move nestedScrollView(…) from sample to new split
